Arreglando pipes y reports

+++ app/models/pipes/pipes.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, DateTime, Boolean
+from sqlalchemy import Column, Integer, String, DateTime, Boolean, Numeric
 from app.models.pipes.pipe_connections import pipe_connections
 from app.models.tanks.tanks_pipes import tank_pipes
 from sqlalchemy.orm import relationship
@@ -6,13 +6,13 @@ from app.db.database import Base
 from geoalchemy2 import Geometry
 from datetime import datetime
 
-class Pipes(Base):
+class Pipes(Base): 
     __tablename__ = "pipes"
     id_pipes= Column(Integer, primary_key=True, index=True)
     material= Column(String(50), nullable=False)
-    diameter= Column(Integer, nullable=False)
+    diameter= Column(Numeric(10, 6), nullable=False)
     status=Column(Boolean, default=True)
-    size= Column(Integer, index=True, nullable=False)
+    size= Column(Numeric(10, 6), index=True, nullable=False)
     installation_date = Column(DateTime, nullable=True)
     coordinates = Column(Geometry(geometry_type="POINT", srid=4326), nullable=False)
     observations=Column(String(100), nullable=True)
